Merge branch 'frasers' into jake

Conflicts:
	Whirlpool/bin/Whirlpool.apk
	Whirlpool/bin/classes.dex
	Whirlpool/bin/classes/com/sinkingduckstudios/whirlpool/R$raw.class
	Whirlpool/bin/classes/com/sinkingduckstudios/whirlpool/environment/Finish.class
	Whirlpool/bin/classes/com/sinkingduckstudios/whirlpool/logic/Level.class
	Whirlpool/bin/resources.ap_
	Whirlpool/gen/com/sinkingduckstudios/whirlpool/R.java
	Whirlpool/src/com/sinkingduckstudios/whirlpool/logic/Level.java

+++ Whirlpool/src/com/sinkingduckstudios/whirlpool/logic/Screen.java
@@ -41,10 +41,10 @@ public class Screen {
 		mCentre.setPoints(width/2, height/2);
 	}
 	//getters for x and y components
-	public int getCentreX(){
+	public float getCentreX(){
 		return mCentre.getX();
 	}
-	public int getCentreY(){
+	public float getCentreY(){
 		return mCentre.getY();
 	}
 }
